[ADD]: Add a ChatGPT script "equivalent" to mine
Of course it can solve this, it's undergrad maths!

It is a little more intuitive than mine - it normalizes probabilities (I print P(the sequence occuring | given k pairs of socks), this prints P(there being k pairs of socks | given our sequence); it can be shown that this is equivalent, since we assume a uniform prior distribution).